+++ lib/GoodsPages/goodsDetail_New.dart
@@ -16,8 +16,7 @@ import '../model/Favorite_button.dart';
 import 'bar_chart.dart';
 
 class GoodsDetailHome extends StatefulWidget {
-
-
+  final NewGoods goods;
   const GoodsDetailHome({Key key, @required this.goods}) : super(key: key);
 
   @override
