+++ lib/ProfilePages/NoticeHome.dart
@@ -1,22 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:newitem_alarm/ProfilePages/NoticeList.dart';
+import 'package:newitem_alarm/ProfilePages/NoticeDetail.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-class Notice extends StatelessWidget {
-  String datetime;
-  String title;
-  String description;
-  Notice({Key key,  this.datetime, this.title, this.description}) : super(key: key);
+class NoticeHome extends StatelessWidget {
   CollectionReference noticeRef = FirebaseFirestore.instance.collection("Notice");
 
   @override
   Widget build(BuildContext context) {
-    List data_list = [];
-    noticeRef.get().then((value) => value.docs.forEach((element) {data_list.add(element.data());}));
-    datetime = data_list[0];
-    title = data_list[1];
-    description = data_list[2];
     return Scaffold(
       body: StreamBuilder(
         stream: noticeRef.snapshots(),
@@ -37,6 +28,9 @@ class Notice extends StatelessWidget {
                   Timestamp tt = document["datetime"];
                   DateTime now = DateTime.now();
                   String formattedDate = DateFormat('yyyy-MM-dd').format(now);
+                  String datetime = formattedDate.toString();
+                  String title = document["Title"].toString();
+                  String description = document["description"].toString();
                   return Padding(
                     padding: const EdgeInsets.all(3.0),
                     child: GestureDetector(
@@ -50,12 +44,12 @@ class Notice extends StatelessWidget {
                               Row(
                                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                                 children: <Widget>[
-                                  Text(document["Title"],
+                                  Text(title,
                                     style:TextStyle(
                                       color: Colors.black, fontSize: 18.0, fontWeight: FontWeight.bold,
                                     ),
                                   ),
-                                  Text(formattedDate.toString(),
+                                  Text(datetime,
                                   style: TextStyle(color:  Colors.grey[600]),
                                   )
                                 ],
@@ -68,7 +62,6 @@ class Notice extends StatelessWidget {
                                   document["description"],
                                   maxLines: 3,
                                   style: TextStyle(color: Colors.black54),
-
                                 ),
                               )
                             ],
@@ -78,14 +71,13 @@ class Notice extends StatelessWidget {
                       onTap: (){
                         Navigator.push(
                           context,
-                          MaterialPageRoute(builder: (context) => NoticeList()),
+                          MaterialPageRoute(builder: (context) => NoticeDetail(datetime: datetime, title: title, description: description)),
                         );
                       },
                     ),
                   );
                 }).toList();
             List<Widget> noticeItem = snapshotList.cast<Widget>();
-
             return CustomScrollView(
               // CustomScrollView는 children이 아닌 slivers를 사용하며, slivers에는 스크롤이 가능한 위젯이나 리스트가 등록가능함
               // CustomScrollView 아래에는
@@ -100,7 +92,6 @@ class Notice extends StatelessWidget {
                           fontWeight: FontWeight.bold, fontSize: 25, color: Colors.black),
                     ),
                   ),
-
                   // floating 설정. SliverAppBar는 스크롤 다운되면 화면 위로 사라짐.
                   // true: 스크롤 업 하면 앱바가 바로 나타남. false: 리스트 최 상단에서 스크롤 업 할 때에만 앱바가 나타남
                   leading: IconButton(
